+++ CHANGELOG.md
@@ -9,6 +9,7 @@ Version 1.5.1 (not released yet)
 - fix wrong aspect ratio in level editor when used with a wide screen
 - use Hor+ FOV scaling method in level editor perspective view
 - make level editor perspective view FOV depend on view aspect ratio
+- add Save, Pack, Lighting, Play, Play (camera) buttons in top bar of main editor window
 
 Version 1.5.0 (released 2020-09-05)
 -----------------------------------